Perfecting the analysis functions one by one...
They look good for now. Nothing major has changed except that I added a function for replacing the NaN values because I was going through that process in a lot of the functions. I also made the Z Score plot functions much neater. Also, I decided to use some of the parameters from the runexperiment scripts as inputs for the analysis functions, so I have to go back to the runexperiment files and make sure I'm saving some variables into the saved mat file that I wasn't before. I'll go ahead and finish the analysis functions for the other experiments asap.